Move DataTransformer from OroFormBundle to PimEnrichBundle
