+++ src/Pim/Bundle/EnrichBundle/Form/Type/EntityIdentifierType.php
@@ -4,11 +4,11 @@ namespace Pim\Bundle\EnrichBundle\Form\Type;
 
 use Doctrine\Common\Persistence\ManagerRegistry;
 use Doctrine\ORM\EntityManager;
-use Oro\Bundle\FormBundle\Form\DataTransformer\ArrayToStringTransformer;
-use Oro\Bundle\FormBundle\Form\DataTransformer\EntitiesToIdsTransformer;
-use Oro\Bundle\FormBundle\Form\DataTransformer\EntityToIdTransformer;
-use Oro\Bundle\FormBundle\Form\EventListener\FixArrayToStringListener;
-use Oro\Bundle\FormBundle\Form\Exception\FormException;
+use Pim\Bundle\EnrichBundle\Form\DataTransformer\ArrayToStringTransformer;
+use Pim\Bundle\EnrichBundle\Form\DataTransformer\EntitiesToIdsTransformer;
+use Pim\Bundle\EnrichBundle\Form\DataTransformer\EntityToIdTransformer;
+use Pim\Bundle\EnrichBundle\Form\Subscriber\FixArrayToStringListener;
+use Pim\Bundle\EnrichBundle\Form\Exception\FormException;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\Options;
